修改字段

+++ bmc-api/src/main/java/com/ai/baas/bmc/api/proferentialprocuct/params/SingleProductInfo.java
@@ -74,7 +74,7 @@ public class SingleProductInfo extends BaseResponse {
 	/**
 	 * 已选择的产品列表,必填
 	 */
-	private List<Integer> productList;
+	private List<Long> productList;
 	
 	/**
 	 * 操作员Id
@@ -145,13 +145,14 @@ public class SingleProductInfo extends BaseResponse {
 	public void setComments(String comments) {
 		this.comments = comments;
 	}
-	public List<Integer> getProductList() {
+	
+	
+	public List<Long> getProductList() {
 		return productList;
 	}
-	public void setProductList(List<Integer> productList) {
+	public void setProductList(List<Long> productList) {
 		this.productList = productList;
 	}
-	
 	public String getOperatorId() {
 		return operatorId;
 	}
